scripts/sdl3/pyproject.toml
[project]
name = "sdl3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "cython>=3.2.1",
    "libclang>=18.1.1",
]
